+++ tests/testthat/test-missing_arg.R
@@ -1,0 +1,5 @@
+test_that("missing_arg works", {
+    lmargs <- formals(lm)
+    class(lmargs) # pairlist
+    expect_true(missing_arg(lmargs$data))
+})